requirements.txt
line-bot-sdk
flask
imdbpy
pyowm